Define and use the concept of a sub-tower
This provides a better basis for stating that the satisfying monad is
part of the original monadic stack.

However, we unfortunately need to enable OverlappingInstances for this
to work.

It has the added advantage of cleaning up the constraints for
MonadLevel_ quite nicely.

Ideally, this would also be used for MonadTower but that leads to
problems with SatisfyConstraint_.